requirements.txt
numpy
opencv-python
pillow
pyyaml
ultralytics
torch
torchvision
matplotlib
pandas
tqdm
scikit-learn
